@@ -13,7 +13,6 @@
   "files": [
     "bin",
     "config",
-    "bin",
     "scripts",
     "template",
     "utils"
@@ -22,24 +21,21 @@
     "react-scripts-ts": "./bin/react-scripts-ts.js"
   },
   "dependencies": {
-    "autoprefixer": "7.1.1",
-    "app-root-path": "^2.0.1",
+    "autoprefixer": "7.1.2",
     "case-sensitive-paths-webpack-plugin": "2.1.1",
     "chalk": "1.1.3",
-    "cli-highlight": "1.1.4",
     "css-loader": "0.28.4",
     "dotenv": "4.0.0",
-    "extract-text-webpack-plugin": "2.1.2",
+    "extract-text-webpack-plugin": "3.0.0",
     "file-loader": "0.11.2",
     "fs-extra": "3.0.1",
-    "html-webpack-plugin": "2.28.0",
-    "jest": "20.0.3",
+    "html-webpack-plugin": "2.29.0",
+    "jest": "20.0.4",
     "object-assign": "4.1.1",
-    "postcss-flexbugs-fixes": "3.0.0",
-    "postcss-loader": "2.0.5",
-    "promise": "7.1.1",
-    "react-dev-utils": "^3.0.1",
-    "react-error-overlay": "^1.0.8",
+    "postcss-flexbugs-fixes": "3.2.0",
+    "postcss-loader": "2.0.6",
+    "promise": "8.0.1",
+    "react-dev-utils": "^4.0.1",
     "style-loader": "0.18.2",
     "ts-jest": "^20.0.7",
     "ts-loader": "^2.2.1",
@@ -48,11 +44,11 @@
     "tslint-react": "^3.0.0",
     "typescript": "~2.4.0",
     "source-map-loader": "^0.2.1",
-    "sw-precache-webpack-plugin": "0.11.3",
-    "url-loader": "0.5.8",
-    "webpack": "2.6.1",
-    "webpack-dev-server": "2.5.0",
-    "webpack-manifest-plugin": "1.1.0",
+    "sw-precache-webpack-plugin": "0.11.4",
+    "url-loader": "0.5.9",
+    "webpack": "3.5.1",
+    "webpack-dev-server": "2.7.1",
+    "webpack-manifest-plugin": "1.2.1",
     "whatwg-fetch": "2.0.3"
   },
   "devDependencies": {
