@@ -7,9 +7,12 @@
  * of patent rights can be found in the PATENTS file in the same directory.
  */
 
-'use strict';
+/* @flow */
 
-var Anser = require('anser');
+import Anser from 'anser';
+import { AllHtmlEntities as Entities } from 'html-entities';
+
+var entities = new Entities();
 
 // Color scheme inspired by https://chriskempson.github.io/base16/css/base16-github.css
 // var base00 = 'ffffff'; // Default Background
@@ -60,8 +63,8 @@ var anserMap = {
   'ansi-white': 'darkgrey',
 };
 
-function ansiHTML(txt) {
-  var arr = new Anser().ansiToJson(txt, {
+function generateAnsiHTML(txt: string): string {
+  var arr = new Anser().ansiToJson(entities.encode(txt), {
     use_classes: true,
   });
 
@@ -102,4 +105,4 @@ function ansiHTML(txt) {
   return result;
 }
 
-module.exports = ansiHTML;
+export default generateAnsiHTML;
