@@ -97,7 +97,7 @@ function formatMessage(message, isError) {
   // from user code generated by WebPack. For more information see
   // https://github.com/facebookincubator/create-react-app/pull/1050
   message = message.replace(
-    /^\s*at\s((?!webpack:).)*:\d+:\d+[\s\)]*(\n|$)/gm,
+    /^\s*at\s((?!webpack:).)*:\d+:\d+[\s)]*(\n|$)/gm,
     ''
   ); // at ... ...:x:y
 
@@ -120,11 +120,6 @@ function formatWebpackMessages(json) {
     // This prevents a confusing ESLint parsing error
     // preceding a much more useful Babel syntax error.
     result.errors = result.errors.filter(isLikelyASyntaxError);
-  }
-  // Only keep the first error. Others are often indicative
-  // of the same problem, but confuse the reader with noise.
-  if (result.errors.length > 1) {
-    result.errors.length = 1;
   }
   return result;
 }
