@@ -11,6 +11,8 @@
 
 const chalk = require('chalk');
 const url = require('url');
+const globalModules = require('global-modules');
+const fs = require('fs');
 
 function printHostingInstructions(
   appPackage,
@@ -121,10 +123,12 @@ function printHostingInstructions(
     );
     console.log('You may serve it with a static server:');
     console.log();
-    if (useYarn) {
-      console.log(`  ${chalk.cyan('yarn')} global add serve`);
-    } else {
-      console.log(`  ${chalk.cyan('npm')} install -g serve`);
+    if (!fs.existsSync(`${globalModules}/serve`)) {
+      if (useYarn) {
+        console.log(`  ${chalk.cyan('yarn')} global add serve`);
+      } else {
+        console.log(`  ${chalk.cyan('npm')} install -g serve`);
+      }
     }
     console.log(`  ${chalk.cyan('serve')} -s ${buildFolder}`);
     console.log();
