@@ -13,22 +13,11 @@ import { parse } from './parser';
 import { map } from './mapper';
 import { unmap } from './unmapper';
 
-type ErrorRecord = {
-  error: Error,
-  unhandledRejection: boolean,
-  contextSize: number,
-  enhancedFrames: StackFrame[],
-};
-type ErrorRecordReference = number;
-const recorded: ErrorRecord[] = [];
-
-let errorsConsumed: ErrorRecordReference = 0;
-
-function consume(
+function getStackFrames(
   error: Error,
   unhandledRejection: boolean = false,
   contextSize: number = 3
-): Promise<ErrorRecordReference | null> {
+): Promise<StackFrame[] | null> {
   const parsedFrames = parse(error);
   let enhancedFramesPromise;
   if (error.__unmap_source) {
@@ -49,32 +38,13 @@ function consume(
     ) {
       return null;
     }
-    enhancedFrames = enhancedFrames.filter(
+    return enhancedFrames.filter(
       ({ functionName }) =>
         functionName == null ||
         functionName.indexOf('__stack_frame_overlay_proxy_console__') === -1
     );
-    recorded[++errorsConsumed] = {
-      error,
-      unhandledRejection,
-      contextSize,
-      enhancedFrames,
-    };
-    return errorsConsumed;
   });
 }
 
-function getErrorRecord(ref: ErrorRecordReference): ErrorRecord {
-  return recorded[ref];
-}
-
-function drain() {
-  // $FlowFixMe
-  const keys = Object.keys(recorded);
-  for (let index = 0; index < keys.length; ++index) {
-    delete recorded[keys[index]];
-  }
-}
-
-export { consume, getErrorRecord, drain };
-export type { ErrorRecordReference };
+export default getStackFrames;
+export { getStackFrames };
